+++ src/handler/semanticTokensHighlights/index.ts
@@ -6,9 +6,12 @@ import BufferSync from '../../model/bufferSync'
 import Highlighter from '../../model/highligher'
 import { ConfigurationChangeEvent, HandlerDelegate } from '../../types'
 import { disposeAll } from '../../util'
-import window from '../../window'
+import { distinct } from '../../util/array'
 import workspace from '../../workspace'
-import SemanticTokensBuffer, { NAMESPACE, SemanticTokensConfig } from './buffer'
+import window from '../../window'
+import FloatFactory from '../../model/floatFactory'
+import SemanticTokensBuffer, { capitalize, HLGROUP_PREFIX, NAMESPACE, SemanticTokensConfig } from './buffer'
+import { positionInRange } from '../../util/position'
 const logger = require('../../util/logger')('semanticTokens')
 const headGroup = 'Statement'
 
@@ -17,30 +20,37 @@ export default class SemanticTokensHighlights {
   private config: SemanticTokensConfig
   private disposables: Disposable[] = []
   private highlighters: BufferSync<SemanticTokensBuffer>
+  private floatFactory: FloatFactory
 
   constructor(private nvim: Neovim, private handler: HandlerDelegate) {
     this.loadConfiguration()
+    this.floatFactory = new FloatFactory(nvim)
     workspace.onDidChangeConfiguration(this.loadConfiguration, this, this.disposables)
     commands.register({
       id: 'semanticTokens.checkCurrent',
       execute: async () => {
-        try {
-          let item = await this.getCurrentItem()
-          item.checkState()
-        } catch (e) {
-          window.showMessage(e.message, 'error')
-          return
-        }
-        window.showMessage('Semantic tokens provider found for current buffer', 'more')
+        await this.showHiglightInfo()
       }
-    }, false, 'check semantic tokens provider for current buffer')
+    }, false, 'show semantic tokens highlight information of current buffer')
+    commands.register({
+      id: 'semanticTokens.refreshCurrent',
+      execute: () => {
+        return this.highlightCurrent()
+      }
+    }, false, 'refresh semantic tokens highlight of current buffer.')
+    commands.register({
+      id: 'semanticTokens.inspect',
+      execute: () => {
+        return this.inspectSemanticToken()
+      }
+    }, false, 'Inspect semantic token information at cursor position.')
     commands.register({
       id: 'semanticTokens.clearCurrent',
       execute: async () => {
         let buf = await nvim.buffer
         buf.clearNamespace(NAMESPACE, 0, -1)
       }
-    }, false, 'clear semantic tokens highlights of current buffer')
+    }, false, 'clear semantic tokens highlight of current buffer')
     commands.register({
       id: 'semanticTokens.clearAll',
       execute: async () => {
@@ -49,12 +59,7 @@ export default class SemanticTokensHighlights {
           buf.clearNamespace(NAMESPACE, 0, -1)
         }
       }
-    }, false, 'clear semantic tokens highlights of all buffers')
-    this.disposables.push({
-      dispose: () => {
-        commands.unregister('semanticTokens.checkCurrentBuffer')
-      }
-    })
+    }, false, 'clear semantic tokens highlight of all buffers')
     this.highlighters = workspace.registerBufferSync(doc => {
       return new SemanticTokensBuffer(this.nvim, doc.bufnr, this.config)
     })
@@ -69,32 +74,56 @@ export default class SemanticTokensHighlights {
   }
 
   private loadConfiguration(e?: ConfigurationChangeEvent): void {
-    if (!e || e.affectsConfiguration('coc.preferences')) {
-      // let con = this.config || {}
-      let config = workspace.getConfiguration('coc.preferences')
-      let enabled = config.get<boolean>('semanticTokensHighlights', true)
-      if (workspace.isVim && !workspace.env.textprop) {
-        enabled = false
-      }
-      if (this.config && enabled != this.config.enabled) {
-        if (this.highlighters) {
-          for (let buf of this.highlighters.items) {
-            buf.setState(enabled)
-          }
-        }
-      }
-      if (!this.config) {
-        this.config = { enabled }
-      } else {
-        this.config.enabled = enabled
+    if (!e || e.affectsConfiguration('semanticTokens')) {
+      let config = workspace.getConfiguration('semanticTokens')
+      this.config = Object.assign(this.config || {}, {
+        highlightGroups: [],
+        filetypes: config.get<string[]>('filetypes', []),
+        highlightPriority: config.get<number>('highlightPriority', 2048),
+        incrementTypes: config.get<string[]>('incrementTypes'),
+        combinedModifiers: config.get<string[]>('combinedModifiers')
+      })
+      if (!e) {
+        this.config.highlightGroups = workspace.env.semanticHighlights
       }
     }
   }
 
-  public async getCurrentItem(): Promise<SemanticTokensBuffer> {
+  public async inspectSemanticToken(): Promise<void> {
+    let item = await this.getCurrentItem()
+    if (!item || !item.enabled) {
+      this.floatFactory.close()
+      return
+    }
+    let position = await window.getCursorPosition()
+    let highlight = item.highlights.find(o => positionInRange(position, o.range) == 0)
+    if (highlight) {
+      let modifiers = highlight.tokenModifiers || []
+      let docs = [{
+        filetype: 'txt',
+        content: `Type: ${highlight.tokenType}\nModifiers: ${modifiers.join(', ')}\nHighlight group: ${highlight.hlGroup || ''}`
+      }]
+      await this.floatFactory.show(docs, {
+        autoHide: true,
+        focusable: true,
+        title: 'Semantic token info',
+        borderhighlight: 'MoreMsg',
+        border: [1, 1, 1, 1]
+      })
+    } else {
+      this.floatFactory.close()
+    }
+  }
+
+  public async fetchHighlightGroups(): Promise<void> {
+    let res = await this.nvim.call('coc#util#semantic_hlgroups') as string[]
+    this.config.highlightGroups = res
+  }
+
+  public async getCurrentItem(): Promise<SemanticTokensBuffer | null> {
     let buf = await this.nvim.buffer
     let highlighter = this.highlighters.getItem(buf.id)
-    if (!highlighter) throw new Error('current buffer not attached')
+    if (!highlighter) null
     return highlighter
   }
 
@@ -102,69 +131,76 @@ export default class SemanticTokensHighlights {
    * Force highlight of current buffer
    */
   public async highlightCurrent(): Promise<void> {
-    let highlighter = await this.getCurrentItem()
-    highlighter.checkState()
-    await highlighter.forceHighlight()
+    let item = await this.getCurrentItem()
+    if (!item || !item.enabled) throw new Error(`Unable to perform semantic highlights for current buffer.`)
+    await this.fetchHighlightGroups()
+    await item.forceHighlight()
   }
 
   /**
    * Show semantic highlight info in temporarily buffer
    */
   public async showHiglightInfo(): Promise<void> {
-    if (!this.config.enabled) throw new Error('Semantic highlights is disabled by configuration.')
-    let item = await this.getCurrentItem()
-    item.checkState()
-    let highlights = item.highlights || []
-    let highlighter = new Highlighter()
+    let buf = await this.nvim.buffer
     let { nvim } = this
+    let item = this.highlighters.getItem(buf.id)
+    if (!item) return nvim.echoError('Document not attached.')
+    let highlighter = new Highlighter()
     nvim.pauseNotification()
-    nvim.command(`vs +setl\\ buftype=nofile __coc_semantic_highlights_${item.bufnr}__`, true)
+    nvim.command(`vs +setl\\ buftype=nofile __coc_semantic_highlights_${buf.id}__`, true)
     nvim.command(`setl bufhidden=wipe noswapfile nobuflisted wrap undolevels=-1`, true)
     nvim.call('bufnr', ['%'], true)
     let res = await nvim.resumeNotification()
-    if (res[1]) throw new Error(`Error on buffer create: ${res[1]}`)
-    let bufnr = res[0][2] as number
     highlighter.addLine('Semantic highlights info', headGroup)
     highlighter.addLine('')
-    highlighter.addLine('The number of semantic tokens: ')
-    highlighter.addText(String(highlights.length), 'Number')
-    highlighter.addLine('')
-    highlighter.addLine('Semantic highlight groups used by current buffer', headGroup)
-    highlighter.addLine('')
-    const groups = [...new Set(highlights.map(({ hlGroup }) => hlGroup))]
-    for (const hlGroup of groups) {
-      highlighter.addTexts([{ text: '-', hlGroup: 'Comment' }, { text: ' ' }, { text: hlGroup, hlGroup }])
+    try {
+      item.checkState()
+      let highlights = item.highlights || []
+      highlighter.addLine('The number of semantic tokens: ')
+      highlighter.addText(String(highlights.length), 'Number')
       highlighter.addLine('')
-    }
-    highlighter.addLine('Tokens types that current Language Server supported:', headGroup)
-    highlighter.addLine('')
-    let doc = workspace.getDocument(item.bufnr)
-    const legend = languages.getLegend(doc.textDocument)
-    if (legend?.tokenTypes.length) {
-      for (const t of [...new Set(legend.tokenTypes)]) {
-        highlighter.addTexts([{ text: '-', hlGroup: 'Comment' }, { text: ' ' }, { text: `CocSem_${t}`, hlGroup: `CocSem_${t}` }])
+      highlighter.addLine('Semantic highlight groups used by current buffer', headGroup)
+      highlighter.addLine('')
+      const groups = distinct(highlights.filter(o => o.hlGroup != null).map(({ hlGroup }) => hlGroup))
+      for (const hlGroup of groups) {
+        highlighter.addTexts([{ text: '-', hlGroup: 'Comment' }, { text: ' ' }, { text: hlGroup, hlGroup }])
         highlighter.addLine('')
       }
-    } else {
-      highlighter.addLine('No token types supported', 'Comment')
-    }
-    highlighter.addLine('Tokens modifiers that current Language Server supported:', headGroup)
-    highlighter.addLine('')
-    if (legend?.tokenModifiers.length) {
-      for (const t of [...new Set(legend.tokenModifiers)]) {
-        highlighter.addTexts([{ text: '-', hlGroup: 'Comment' }, { text: ' ' }, { text: `CocSem_${t}`, hlGroup: `CocSem_${t}` }])
-        highlighter.addLine('')
+      highlighter.addLine('Tokens types that current Language Server supported:', headGroup)
+      highlighter.addLine('')
+      let doc = workspace.getDocument(item.bufnr)
+      const legend = languages.getLegend(doc.textDocument)
+      if (legend?.tokenTypes.length) {
+        for (const t of [...new Set(legend.tokenTypes)]) {
+          let text = HLGROUP_PREFIX + capitalize(t)
+          highlighter.addTexts([{ text: '-', hlGroup: 'Comment' }, { text: ' ' }, { text, hlGroup: text }])
+          highlighter.addLine('')
+        }
+      } else {
+        highlighter.addLine('No token types supported', 'Comment')
       }
-    } else {
-      highlighter.addLine('No token modifiers supported', 'Comment')
+      highlighter.addLine('Tokens modifiers that current Language Server supported:', headGroup)
+      highlighter.addLine('')
+      if (legend?.tokenModifiers.length) {
+        for (const t of [...new Set(legend.tokenModifiers)]) {
+          let text = HLGROUP_PREFIX + capitalize(t)
+          highlighter.addTexts([{ text: '-', hlGroup: 'Comment' }, { text: ' ' }, { text, hlGroup: text }])
+          highlighter.addLine('')
+        }
+      } else {
+        highlighter.addLine('No token modifiers exists', 'Comment')
+      }
+    } catch (e) {
+      highlighter.addLine(e.message, 'Error')
     }
     nvim.pauseNotification()
+    let bufnr = res[0][2] as number
     highlighter.render(nvim.createBuffer(bufnr))
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    nvim.resumeNotification(false, true)
+    void nvim.resumeNotification(false, true)
   }
 
   public dispose(): void {
+    this.floatFactory.dispose()
     this.highlighters.dispose()
     disposeAll(this.disposables)
   }